# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BookWorm API
  description: |
    Full-stack online bookstore REST API.
    
    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    
    ```
    Authorization: Bearer <token>
    ```
    
    ## Getting Started
    1. Register at `POST /auth/register` or login at `POST /auth/login`
    2. Copy the returned `token`
    3. Click "Authorize" above and paste: `Bearer <token>`
  version: 1.0.0
  contact:
    name: BookWorm API Support
servers:
  - url: http://localhost:5000/api/v1
    description: Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        is_guest:
          type: boolean
        gift_points:
          type: integer
        created_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /users/profile:
    get:
      tags: [Users]
      summary: Get current user profile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [Users]
      summary: Update user profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                email:
                  type: string
                phone_number:
                  type: string
      responses:
        '200':
          description: Profile updated
  /users/order-history:
    get:
      tags: [Users]
      summary: Get order history with pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Order history
  /users/wishlist:
    get:
      tags: [Users]
      summary: Get user wishlist
      responses:
        '200':
          description: Wishlist items
    post:
      tags: [Users]
      summary: Add book to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [book_id]
              properties:
                book_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Added to wishlist
  /users/wishlist/{bookId}:
    delete:
      tags: [Users]
      summary: Remove book from wishlist
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Removed from wishlist
...
  /books/{bookId}/reviews:
    get:
      tags: [Books]
      summary: Get book reviews
      security: []
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Book reviews

  # ── Categories ──────────────────────────────────────────────────────
...
  /payments/initiate:
    post:
      tags: [Payments]
      summary: Initiate payment for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order_id, payment_method]
              properties:
                order_id:
                  type: string
                  format: uuid
                payment_method:
                  type: string
                  enum: [credit_card, debit_card, upi, wallet]
                card_last_four:
                  type: string
                  minLength: 4
                  maxLength: 4
      responses:
        '201':
          description: Payment initiated
  /payments/confirm:
    post:
      tags: [Payments]
      summary: Confirm payment (mock gateway)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payment_id]
              properties:
                payment_id:
                  type: string
                  format: uuid
                simulate_failure:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Payment result
  /payments/refund:
    post:
      tags: [Payments]
      summary: Process a refund
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payment_id]
              properties:
                payment_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Refund processed
  /payments/{paymentId}/status:
    get:
      tags: [Payments]
      summary: Get payment status
      parameters:
        - in: path
          name: paymentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment status

  # ── Reviews ─────────────────────────────────────────────────────────
  /reviews:
    post:
      tags: [Reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [book_id, rating]
              properties:
                book_id:
                  type: string
                  format: uuid
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                review_text:
                  type: string
      responses:
        '201':
          description: Review submitted
  /reviews/{reviewId}:
    put:
      tags: [Reviews]
      summary: Update a review
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review updated
    delete:
      tags: [Reviews]
      summary: Delete a review
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review deleted

  # ── Recommendations ─────────────────────────────────────────────────
...